+++ juniper-from-schema-build-tests/basic/build.rs
@@ -1,0 +1,14 @@
+fn main() {
+    juniper_from_schema_build::compile_schema_literal(
+        r#"
+        schema {
+            query: Query
+        }
+
+        type Query {
+            ping: Boolean!
+        }
+    "#,
+    )
+    .unwrap();
+}